Generalised caching, more documentation, and other set-up things
This change adds a PSR-6 compatible caching system, instead of
only using Redis. This makes it easier to run locally (without
installing Redis).

Also extends the local-development setup instructions, perhaps too
verbosely, making it clearer what needs to be done to get things
up and running. Composer will now create a default .env file.

The Less cache is moved to the top-level 'cache' directory, becuase
this is where the cache is when Redis isn't used and because /src
seems like an odd place for volatile files.

The /login route is removed because it wasn't used anyway.