Enable use on all Wikipedias
This change updates CopyPatrol to work on all language
Wikipedias that are supported by EranBot. Primarily it
focusses on getting things working with French (for
which translations are encluded here) but with the
exception of the WikiProject prefix cleanup in index.js
there's now nothing here specific to any one language.

Upgrading to this will require a couple of changes in
the .env file, and running `composer install`.

Bug: https://phabricator.wikimedia.org/T145436