Fix Leaderboard language-specific links
This updates the Leaderboard controller to follow the same pattern
as the other controllers and inherit from CopyPatrol. This makes
it simpler to retrieve the current wiki langugage in the same way
that it's done elsewhere (and so is hopefully less confusing).

Some of the Twig variable usages were also simplified to the dot
syntax rather than the array syntax.

Bug: https://phabricator.wikimedia.org/T150928