Add a dropdown for switching between Wikipedias
This changes to using the Bootstrap navbar for the main menu, so
that it's easier to incorporate a dropdown with which to switch
from English to French or vice versa. This switch will to to the
corresponding other-language page (so if you're on the leaderboard
and switch you'll be on the other language's leaderboard).

Bug: https://phabricator.wikimedia.org/T150956