Only show the WikiProject field when required
This change hides the WikiProject search field and results column
if there are no WikiProjects in the database for the Wikipedia
language in use. The other results columns are made wider to fill
the gap.

Half of T152173 is fixed as well, because it's easier to test
this new functionality against a language (pt) that doesn't
have WikiProjects.

Composer is also updated, just for fun.

Bug: https://phabricator.wikimedia.org/T152134
Bug: https://phabricator.wikimedia.org/T152173