Validate the wiki language in the route pattern (#45)
Before, if an invalid language was requested (but one that still
passed the wikiLang regex, e.g. qq) the app would use it regardless.
This change dynamically constructs the route regex so that it checks
the incoming language against the list of supported languages. This
way a 404 will be thrown as normal if it doesn't match.

The whole route code layout is changed a bit, to group all the
wikiLang routes together and make the others a little more thematically
located. There's some duplication in the now not-grouped review
routes, becuase it seemed clearer to not have an extra level of
grouping.

Bug: https://phabricator.wikimedia.org/T152134