Add Dockerfile, development docker-compose
This commit adds a Dockerfile for building a reproducible image of CopyPatrol
based on Toolforge PHP 7.4 images. The Dockerfile mostly emulates the
Toolforge environment, with slight modifications in development mode to make
development easier (such as enabling XDebug in development mode).

The `docker-compose.dev.yml` file ensures that CopyPatrol is started with
additional services, such as Redis, and mounts, such as the main program mount
on the public user directory. It also ensures that port 9000/9003 are made
accessible within the container to allow remote XDebug step debugging.

Note that the Docker Compose file binds to port 80 for the webserver. As the
Twig templates are written now, ports are not properly handled when linking
to server resources.