Port Travis to GitHub Actions
Builds on `travis-ci.org` ended back in June 2021. This means the Travis
CI has not been providing builds for this repo since then. Switching to
GitHub Actions has been the course of action for pywikibot, and GitHub
Actions supports an unlimited amount of free minutes for public projects,
so this is a switch to GitHub Actions to get continuous integration back up.

From https://docs.github.com/en/billing/managing-billing-for-github-actions/about-billing-for-github-actions:
> GitHub Actions usage is free for standard GitHub-hosted runners in public
> repositories, and for self-hosted runners.