Convert to Symfony and use ToolforgeBundle
This is more or less a rewrite, but visually nothing should have
changed. Moving to Symfony and using ToolforgeBundle makes the app more
maintainable, especially for Community Tech (the maintaining team).

All controller actions are now in a single class, AppController. From
here, we build Record objects that get pass to the views. Just as with
before, there are separate classes to communicate with the db and APIs,
now called CopyPatrolRepository and WikiRepository.

Summary of other changes:
* Add .editorconfig
* Use .env.local in local envs, with .env as the template
* Add .nvmrc
* Covert JS to ES6 and use Webpack Encore
* Remove old browser-prefixed rules from the LESS
* Rename status values to work with new bot/schema
* Update all PHP dependencies, remove those we don't need
* Remove code that now lives in ToolforgeBundle
* Remove unused messages from i18n
* Use NPM releases of bootstrap and jquery, and only include the
  parts that we need.
* Remove Select2 JS library as we're not using it anymore
* Do all validations server-side and not just in the JS
* Add caching to main DB queries and HTTP requests
* Convert the data structure for the view in to a Record model
* Fetch WikiProjects at the same time as redlink status of page titles
* Add stylelint and fix violations
* Add integration tests
* Other various code cleanup

This rewrite does NOT check against the user allowlist and auto-mark
records for users who are on it. That job is going to be moved to the
bot. See T341726.

Bug: T340600