Record: type cast page_namespace into int; make submission_id a string
The version and/or configuration of the MySQL client can apparently mean
we get a string instead of an integer, causing the page titles to be
displayed incorrectly.

For good measure, we type cast all getters that should return an int.

The submission_id meanwhile can either be an int or a string, depending
on whether it's from the new Turnitin API or the old one. We cast to
string since we don't need the integer type for this field.

Also update the 'record-url-text' message, which no longer shows the
number of words due to a limitation with the new Turnitin API.