Add links to delete page, undo edit and revdel edit when applicable
There are FIXMEs where we are hard-coding enwiki-specific values.
Later, we might consider making an on-wiki interface page where admins
can cusomtize the links to suit their workflows.

Also change en.json to use tabs insteads of spaces, which makes it
consistent with (most) other JSON files in this repo.